Update to changeset: 534:eb18c78ef811 for libtcod
* Changed the code source of libtcod
* Changed build.rs to not use deprecated gcc::Config

+++ tcod_sys/libtcod/src/fov_permissive2.c
@@ -24,12 +24,14 @@
 * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
+#include <fov.h>
 
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
-#include "libtcod.h"
-#include "libtcod_int.h"
+
+#include <libtcod_int.h>
+#include <libtcod_utility.h>
 
 /* The size of each square in units */
 #define STEP_SIZE 16
@@ -168,7 +170,7 @@ static void visit_coords(map_t *m,int startX, int startY, int x, int y, int dx, 
 		/* view split */
 		int offset=startX+x*dx/STEP_SIZE + (startY+y*dy/STEP_SIZE)*m->width;
 		view_t *shallower_view= & views[offset];
-		int view_index=current_view - (view_t **)TCOD_list_begin(active_views);
+		int view_index=(int)(current_view - (view_t **)TCOD_list_begin(active_views));
 		view_t **shallower_view_it;
 		view_t **steeper_view_it;
 		*shallower_view=**current_view;
