+++ tcod_sys/libtcod/src/lex_c.c
@@ -24,17 +24,18 @@
 * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
+#include <lex.h>
+
 #include <ctype.h>
 #include <string.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include "libtcod.h"
 
 #define MAX_JAVADOC_COMMENT_SIZE 16384
 
 /* damn ANSI C does not know strdup, strcasecmp, strncasecmp */
 char *TCOD_strdup(const char *s) {
-	uint32 l=strlen(s)+1;
+	size_t l=strlen(s)+1;
 	char *ret=malloc(sizeof(char)*l);
 	memcpy(ret,s,sizeof(char)*l);
 	return ret;
@@ -331,7 +332,7 @@ int TCOD_lex_get_space(TCOD_lex_t *lex)
 		break;
 	}
 	if ( (lex->flags & TCOD_LEX_FLAG_TOKENIZE_COMMENTS) && startPos && lex->pos > startPos ) {
-		int len = lex->pos - startPos;
+		int len = (int)(lex->pos - startPos);
 		allocate_tok(lex, len+1);
 		strncpy(lex->tok,startPos,len);
 		lex->tok[len]=0;
